добавила выделение выбранного элемента. пока только радио
